% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_counts.R
\name{cat_counts}
\alias{cat_counts}
\title{Count observations or units per node}
\usage{
cat_counts(spec)
}
\arguments{
\item{spec}{A cat_spec or labeled cat_spec object}
}
\value{
A tibble with counts per node
}
\description{
Count observations or units per node
}
