% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_counts.R
\name{cat_counts}
\alias{cat_counts}
\title{Count observations or units per node}
\usage{
cat_counts(spec)
}
\arguments{
\item{spec}{A cat_spec or labeled cat_spec object}
}
\value{
A tibble with counts per node
}
\description{
Count observations or units per node
}
\examples{
df <- data.frame(
  id   = rep(1:4, each = 3),
  year = rep(2018:2020, 4),
  g    = c(rep(2019, 6), rep(Inf, 6))
)
spec <- cat_spec(df, id = "id", time = "year", g = "g")
cat_counts(spec)
}
